apis/v1: Split Prometheus types in their own file
